requirements.txt
asgiref==3.8.1
Django==5.1.5
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
django-mail-templated==2.6.5
